Run variant extras before build purge
Pull the inclusion of variant extras further up, to have build
dependencies available during that phase. This allows to build
additional PHP extensions by defining the run scripts in *-extras files.

To have the list of manually installed packages persistent between the
now split RUN commands, the list of packages is written to a file
instead of having it available as environmental variable.