Add apache variant with Xdebug enabled
Add a new variant with Xdebug extension enabled for remote debugging.
As the `variant` variable used inside the update.sh script is a direct
reference to the docker image used as base for the Dockerfile, a new
variable to define a differing base-image is introduced.

This variant is intended for development purposes only!